+++ packages/ckeditor5-ckbox/tests/ckboximageedit/ckboximageeditcommand.js
@@ -372,79 +372,6 @@ describe( 'CKBoxImageEditCommand', () => {
 				sinon.assert.callCount( respondSpy, 4 );
 			} );
 
-			it( 'should abort the request if cannot determine a response due to network error', () => {
-				const spy = sinon.spy( command.controller.signal, 'throwIfAborted' );
-
-				sinonXHR.respondWith( 'GET', CKBOX_API_URL + '/assets/image-id1', xhr => {
-					expect( xhr.requestHeaders ).to.be.an( 'object' );
-					expect( xhr.requestHeaders ).to.contain.property( 'Authorization', jwtToken );
-					expect( xhr.requestHeaders ).to.contain.property( 'CKBox-Version', 'CKEditor 5' );
-
-					return xhr.error();
-				} );
-
-				const dataMock = {
-					id: 'image-id1',
-					extension: 'png',
-					metadata: {
-						width: 100,
-						height: 100
-					},
-					name: 'image1',
-					imageUrls: {
-						100: 'https://example.com/workspace1/assets/image-id1/images/100.webp',
-						default: 'https://example.com/workspace1/assets/image-id1/images/100.png'
-					},
-					url: 'https://example.com/workspace1/assets/image-id1/file'
-				};
-
-				return command._getAssetStatusFromServer( dataMock )
-					.then( () => {
-						throw new Error( 'Expected to be rejected.' );
-					}, () => {
-						sinon.assert.calledOnce( spy );
-					} );
-			} );
-
-			it( 'should throw an error on abort', () => {
-				sinonXHR.respondWith( 'GET', CKBOX_API_URL + '/assets/image-id1', [
-					200,
-					{ 'Content-Type': 'application/json' },
-					JSON.stringify( {
-						metadata: {
-							metadataProcessingStatus: 'queued'
-						}
-					} )
-				] );
-
-				const dataMock = {
-					id: 'image-id1',
-					extension: 'png',
-					metadata: {
-						width: 100,
-						height: 100
-					},
-					name: 'image1',
-					imageUrls: {
-						100: 'https://example.com/workspace1/assets/image-id1/images/100.webp',
-						default: 'https://example.com/workspace1/assets/image-id1/images/100.png'
-					},
-					url: 'https://example.com/workspace1/assets/image-id1/file'
-				};
-
-				const promise = command._getAssetStatusFromServer( dataMock )
-					.then( () => {
-						throw new Error( 'Promise should throw.' );
-					} )
-					.catch( () => {
-						expect( command.controller.signal.aborted ).to.equal( true );
-					} );
-
-				command.abort();
-
-				return promise;
-			} );
-
 			it( 'should reject if fetching asset\\s status ended with the authorization error', () => {
 				sinonXHR.respondWith( 'GET', CKBOX_API_URL + '/assets/image-id1', [
 					401,
